Added support for executing MGFX under mono from the Pipeline tools
Because we need to run MGFX under mono on
Linux and MacOS we need to conditionally alter
the Process parameters. Rather than doing this
in the common code we call out to the IView
to have it create the Process Object for us.

Under Xwt we can run under either mono or
.net , but at least we keep the platform
specific code out of the Common code.