Removed IndependentAxes dead zone calculation
Thumbsticks have to always be considered as "circular", even if they are on independent axes on the hardware/driver side. Not doing so may expose clamping on the opposite axis when one axis if to its min or max value. This would results in a "sticky feel" when pushing a thumbstick to full X/-X or Y/-Y.