Add breaking test: An empty list of a custom class containing a custom enum member. Test fails with `ContentLoadException : Content reader could not be found for CustomEnum type.`
I believe the issue is that when compiling the xml, the custom class is not encountered because there are no items in the list. This means it does not scrape the custom class and generate a EnumReader<CustomEnum> to be added to the ContentTypeReader list at the start of the Xnb file.