Clean up some code that potentially was reflected.
After an incident with a contributor and auditing their contributions we
have a strong suspicion that some bits of their code came from either a
copyrighted source or from a reflection tool.

While we do not have definitive proof of this and cannot be certain of
the code's origin, we feel it is just better for the project to remove
those chunks of code.

 The code in question is in the following functions:

   - BoundingFrustum.Intersects(ref Ray ray, out float? result)
   - Vector4.Transform(ref Vector2 value, ref Quaternion rotation, out Vector4 result)
   - Vector4.Transform(ref Vector3 value, ref Quaternion rotation, out Vector4 result)
   - Vector4.Transform(ref Vector4 value, ref Quaternion rotation, out Vector4 result)

We hope that someone in the community can contribute their own implementations
to fix these holes soon (of course without reading or reusing any of the code removed in this PR).

We will also follow up with an update to the existing CONTRIBUTING.md to help further clarify what we require from them.