Updated Xamarin Studio Shared project to include a Content.mgcb this is because it now works with the Content Pipleine on Mac.
Added a Stand Alone Shared Project for Visual Studio to Mirror the
one in Xamarin studio.

Note : For some reason even though we have the Content.mgcb added as
a item in the shared project Visual Studio does NOT show it. However
it DOES still pick it up and compile it via MGCB.exe