Change the XS ProjectTemplates to use the Local Installed Assemblies.
Using the Nuget packages for the templates was a hack to work around
issues with Assembly resolution in Xamarin Studios/MonoDevelop.
With the changes to the MonoGameMSBuildImports class we can now
correctly inject the HintPath to pick up the local assemblies. This
will allow develop builds to actually reference the installed assemblies
rather than an out of date nuget package.

We do this by having a new msbuild .props file

	MonoGame.Common.props

This file is responsible for finding the correct Instllation path,
all the VS templates have been updated to use this new file and the
old settings to <MonoGameInstallDirectory> have been removed.
The XS templates will also get this new import when they are first loaded.

NOTE : the Import MUST be the first item in the project. This is to the
directory is resolved BEFORE the assemblies are loaded.

Updated Android Templates to use OpenTK-1.0