Fixes #4538 Fullscreen rendering on MacOS
A previous commit b668e0ae changes when the ClientRectangle
was being calculated in the OpenTK DesktopGL system. This appers
to be happening too late on MacOS as a result the Viewport no longer
gives the full desktop resolution when in fullscreen mode.

This commit hacks around this issue by running the old code if we
are not on Linux which is the platform commit b668e0ae was designed
to fix.