Fixes a potential crash when didMoveToWindow is called
This fixes a potential crash when the framebuffer is initialized without a depthbuffer and didMoveToWindow is called (triggered when coming back to the iOSGameView from another View).
In that case the previous code would always try to recreate the framebuffer (since `_depthbuffer == 0`) even if their is one currently active and bound, causing a faulty framebuffer status.