Remove restriction on HardwareModeSwitch
The HardwareModeSwitch property controls if the game window
goes into hard full screen or into a 'soft' fullscreen (which
is basically a fullscreen borderless form).

There is a use case where the developer might want to allow
the user the ability to switch between the two. Plus I cannot
see any hardware reason why this cannot be changed at runtime.