Fixed OpenAL CallingConvention on Function Pointers #5100
The function pointers for OpenAL effects/filters were using
the platform default calling convention. On Mac and Linux
this is Cdecl, but on Windows it is StdCall.

OpenAL soft is built on windows using cdecl. As a result the
calls to these methods would either crash or produce unexpected
behaviour.