Filter Windows native messages
This filters messages above 32 so specifically the WM_TIMER is filtered
out, this event would prevent us from reading the WM_QUIT message and
exiting the application. More explanation and discussion in #3995
I also made it ignore the peeking in DEBUG mode since it's only necessary
when testing. Not sure if the sleeping is still necessary with this...