Make VertexChannel only accept structs
This is a point of discussion maybe because it does not follow the XNA  API. However it does not make sense
to have a reference type for vertex data and when a reference type is used and no vertex data is passed
MeshBuilder.FinishMesh in XNA crashes with a NullReferenceException. So I think that's just a bug they
overlooked. If people don't like this change we can just check if the passed type
in MeshBuilder.CreateVertexChannel is a non-nullable value type like this:

```
if (default(T) == null)
    throw new ArgumentException("T must be a non-nullable value type!", "T");
```